fix is_var test

+++ test/unit/math/rev/meta/is_var_test.cpp
@@ -7,8 +7,8 @@ TEST(MetaTraitsRevScal, is_var) {
   using stan::math::var_value;
   EXPECT_TRUE(is_var<stan::math::var>::value);
   EXPECT_TRUE((is_var<stan::math::var_value<float>>::value));
-  EXPECT_TRUE((is_var<stan::math::var_value<int>>::value));
+  EXPECT_TRUE((is_var<stan::math::var_value<long double>>::value));
   EXPECT_FALSE(is_var<stan::math::vari>::value);
   EXPECT_FALSE((is_var<double>::value));
-  EXPECT_FALSE((is_var<stan::math::vari_value<int>>::value));
+  EXPECT_FALSE((is_var<stan::math::vari_value<double>>::value));
 }
